+++ src/bl64-msg-setup.bash
@@ -154,6 +154,8 @@ function bl64_msg_set_theme() {
 #
 # * Will also setup the theme
 # * If no theme is provided then the STD is used (ansi or ascii)
+# * If no output is provided then the default is used.
+# * Default output is ANSI if the script is interactive, otherwise ASCII
 #
 # Arguments:
 #   $1: output type. One of BL64_MSG_OUTPUT_*. Default: BL64_MSG_OUTPUT_ANSI
@@ -167,10 +169,16 @@ function bl64_msg_set_theme() {
 #######################################
 function bl64_msg_set_output() {
   _bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
-  local output="${1:-${BL64_MSG_OUTPUT_ANSI}}"
+  local output="${1:-}"
   local theme="${2:-${BL64_VAR_DEFAULT}}"
 
-  bl64_check_parameter 'output' || return $?
+  if bl64_lib_var_is_default "$output"; then
+    if [[ "$-" == *i* ]]; then
+      output="$BL64_MSG_OUTPUT_ANSI"
+    else
+      output="$BL64_MSG_OUTPUT_ASCII"
+    fi
+  fi
 
   case "$output" in
   "$BL64_MSG_OUTPUT_ASCII")